+++ packages/core/src/session/stateless/refreshSession.ts
@@ -6,8 +6,8 @@ import type { InternalStatelessContext, Session, User } from "@/@types/index.ts"
 export const refreshSession = <DefaultUser extends User = User>({ ctx, cookies, cookieManager }: InternalStatelessContext) => {
     const { logger, jose, jwtManager: jwt, identity, sessionConfig } = ctx
 
-    const maxAge = sessionConfig?.jwt?.maxAge ?? 60 * 60 * 24 * 15
-    const strategy = sessionConfig?.jwt?.expirationStrategy ?? "absolute"
+    const maxAge = sessionConfig?.maxAge ?? sessionConfig?.jwt?.maxAge ?? 60 * 60 * 24 * 15
+    const strategy = sessionConfig?.expirationStrategy ?? sessionConfig?.jwt?.expirationStrategy ?? "absolute"
 
     return async (
         headers: Headers,